requirements.txt
python-dotenv
pandas 
numpy
scikit-learn
pymongo
certifi
pymongo[srv]==3.6
pymongo
pyYAML
dill

# -e .  